// voro-lp-front/types/Enums/proposalStatusEnum.enum.ts
export enum ProposalStatusEnum {
  Unspecified = 0,       // Não especificado
  Approved = 100,        // Aprovada
  Rejected = 200,        // Rejeitada
  Pending = 300,         // Pendente
  UnderReview = 400,     // Em análise
  Cancelled = 500,       // Cancelada
  OnHold = 600           // Em espera
}